feat!: remove v1 route convention codepaths
